+++ tools/db_crashtest.py
@@ -259,6 +259,10 @@ default_params = {
     "use_dynamic_delay": lambda: random.choice([0, 1, 1, 1]),
     "allow_wbm_stalls": lambda: random.randint(0, 1),
     "start_delay_percent": lambda: random.randint(0, 99),
+    "enable_speedb_features": lambda: random.randint(0, 1),
+    "total_ram_size": lambda: random.choice([512 * 1024 * 1024, 1024 * 1024 * 1024]),
+    "max_background_jobs": lambda: random.choice([4, 8]),
+    "crash_test": 1,
 }
 
 _TEST_DIR_ENV_VAR = "TEST_TMPDIR"